+++ src/telnyx/types/calls/__init__.py
@@ -89,8 +89,11 @@ from .action_switch_supervisor_role_params import ActionSwitchSupervisorRolePara
 from .action_start_noise_suppression_params import (
     ActionStartNoiseSuppressionParams as ActionStartNoiseSuppressionParams,
 )
-from .transcription_engine_xai_config_param import (
-    TranscriptionEngineXaiConfigParam as TranscriptionEngineXaiConfigParam,
+from .action_stop_conversation_relay_params import (
+    ActionStopConversationRelayParams as ActionStopConversationRelayParams,
+)
+from .action_start_conversation_relay_params import (
+    ActionStartConversationRelayParams as ActionStartConversationRelayParams,
 )
 from .action_stop_noise_suppression_response import (
     ActionStopNoiseSuppressionResponse as ActionStopNoiseSuppressionResponse,
@@ -104,8 +107,14 @@ from .action_add_ai_assistant_messages_params import (
 from .action_start_noise_suppression_response import (
     ActionStartNoiseSuppressionResponse as ActionStartNoiseSuppressionResponse,
 )
+from .action_stop_conversation_relay_response import (
+    ActionStopConversationRelayResponse as ActionStopConversationRelayResponse,
+)
 from .transcription_engine_azure_config_param import (
     TranscriptionEngineAzureConfigParam as TranscriptionEngineAzureConfigParam,
+)
+from .action_start_conversation_relay_response import (
+    ActionStartConversationRelayResponse as ActionStartConversationRelayResponse,
 )
 from .transcription_engine_google_config_param import (
     TranscriptionEngineGoogleConfigParam as TranscriptionEngineGoogleConfigParam,
@@ -115,9 +124,6 @@ from .transcription_engine_telnyx_config_param import (
 )
 from .action_add_ai_assistant_messages_response import (
     ActionAddAIAssistantMessagesResponse as ActionAddAIAssistantMessagesResponse,
-)
-from .transcription_engine_assemblyai_config_param import (
-    TranscriptionEngineAssemblyaiConfigParam as TranscriptionEngineAssemblyaiConfigParam,
 )
 from .call_control_command_result_with_conversation_id import (
     CallControlCommandResultWithConversationID as CallControlCommandResultWithConversationID,