+++ tests/api_resources/test_wireguard_interfaces.py
@@ -10,7 +10,7 @@ import pytest
 from telnyx import Telnyx, AsyncTelnyx
 from tests.utils import assert_matches_type
 from telnyx.types import (
-    WireguardInterfaceRead,
+    WireguardInterfaceListResponse,
     WireguardInterfaceCreateResponse,
     WireguardInterfaceDeleteResponse,
     WireguardInterfaceRetrieveResponse,
@@ -114,7 +114,9 @@ class TestWireguardInterfaces:
     @parametrize
     def test_method_list(self, client: Telnyx) -> None:
         wireguard_interface = client.wireguard_interfaces.list()
-        assert_matches_type(SyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            SyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -124,7 +126,9 @@ class TestWireguardInterfaces:
             page_number=0,
             page_size=0,
         )
-        assert_matches_type(SyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            SyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -134,7 +138,9 @@ class TestWireguardInterfaces:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         wireguard_interface = response.parse()
-        assert_matches_type(SyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            SyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -145,7 +151,7 @@ class TestWireguardInterfaces:
 
             wireguard_interface = response.parse()
             assert_matches_type(
-                SyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"]
+                SyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
             )
 
         assert cast(Any, response.is_closed) is True
@@ -289,7 +295,9 @@ class TestAsyncWireguardInterfaces:
     @parametrize
     async def test_method_list(self, async_client: AsyncTelnyx) -> None:
         wireguard_interface = await async_client.wireguard_interfaces.list()
-        assert_matches_type(AsyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            AsyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -299,7 +307,9 @@ class TestAsyncWireguardInterfaces:
             page_number=0,
             page_size=0,
         )
-        assert_matches_type(AsyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            AsyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -309,7 +319,9 @@ class TestAsyncWireguardInterfaces:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         wireguard_interface = await response.parse()
-        assert_matches_type(AsyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"])
+        assert_matches_type(
+            AsyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
+        )
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
@@ -320,7 +332,7 @@ class TestAsyncWireguardInterfaces:
 
             wireguard_interface = await response.parse()
             assert_matches_type(
-                AsyncDefaultFlatPagination[WireguardInterfaceRead], wireguard_interface, path=["response"]
+                AsyncDefaultFlatPagination[WireguardInterfaceListResponse], wireguard_interface, path=["response"]
             )
 
         assert cast(Any, response.is_closed) is True